# [SPARK-25354][SQL] Reduce unneeded operation in nextKeyValue process of parquet vectorized record reader