## What changes were proposed in this pull request?

this PR do following in VectorizedParquetRecordReader class:

1. In `nextKeyValue()`, remove calling to `resultBatch()`. Columnar batch initializition by `initBatch()` is moved to `nextBatch()`.

2. In `nextBatch()`, put `if (rowsReturned >= totalRowCount) return false;` before columnVectors resetting. When condition is true, columnVectors resetting need not execute. 

3. In `nextBatch()`, call checkEndOfRowGroup only when `rowsReturned == totalCountLoadedSoFar` is true instead of calling every time.

4. In `checkEndOfRowGroup()`, save columns of requestedSchema for next use instead of reading every time. 

From analysis of JMC tool, we found that frequencies of calling to `nextKeyValue()` is very high, performance gain may be got from these changes.



## How was this patch tested?

1. Existing Unit Tests
2. A test of running 4885 spark applications.