@SongYadong 2018-09-07
In my test, total executing time of 4885 spark applications:
origin: 447213 seconds
after : 444584 seconds
time saved : about 0.6%