@dongjoon-hyun 2018-09-10
@SongYadong . `0.6%` seems to be too small to do this. Do you have another benchmark result to show the benefit more clearly?