@SongYadong 2018-09-12
Sounds reasonable. I will close this PR. Thank you!